pyproject.toml
[project]
name = "mcps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "attrs==25.4.0",
    "authlib==1.6.5",
    "bcrypt==5.0.0",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "colorama==0.4.6",
    "cryptography==46.0.3",
    "cyclopts==3.24.0",
    "distro==1.9.0",
    "dnspython==2.8.0",
    "docstring-parser==0.17.0",
    "docutils==0.22.2",
    "email-validator==2.3.0",
    "exceptiongroup==1.3.0",
    "fastapi==0.119.0",
    "fastmcp==2.12.4",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "idna==3.11",
    "invoke==2.2.1",
    "isodate==0.7.2",
    "jiter==0.11.1",
    "jsonschema==4.25.1",
    "jsonschema-path==0.3.4",
    "jsonschema-specifications==2025.9.1",
    "lazy-object-proxy==1.12.0",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mcp[cli]==1.18.0",
    "mdurl==0.1.2",
    "more-itertools==10.8.0",
    "openai==2.5.0",
    "openapi-core==0.19.5",
    "openapi-pydantic==0.5.1",
    "openapi-schema-validator==0.6.3",
    "openapi-spec-validator==0.7.2",
    "paramiko==4.0.0",
    "parse==1.20.2",
    "pathable==0.4.4",
    "pycparser==2.23",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pydantic-settings==2.11.0",
    "pygments==2.19.2",
    "pyjwt>=2.10.1",
    "pynacl==1.6.0",
    "pyperclip==1.11.0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pywin32==311",
    "pyyaml==6.0.3",
    "referencing==0.36.2",
    "requests==2.32.5",
    "rfc3339-validator==0.1.4",
    "rich==14.2.0",
    "rich-rst==1.3.2",
    "rpds-py==0.27.1",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sse-starlette==3.0.2",
    "starlette==0.48.0",
    "tqdm==4.67.1",
    "typer==0.19.2",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "werkzeug==3.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp_server*"]
